NetworkService implemented
