--- src/main/java/de/nuro/rawlearning/RawlearningApplication.java
@@ -1,12 +1,0 @@
-package de.nuro.rawlearning;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RawlearningApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RawlearningApplication.class, args);
-	}
-}
