+++ src/main/resources/META-INF/resources/frontend/src/view/foto-view.html
@@ -14,8 +14,7 @@
             }
         </style>
 <vaadin-vertical-layout id="main-layout" style="width: 100%; height: 100%;">
- <span id="zahlSpan">Text</span>
- <vaadin-button id="init">init Network</vaadin-button>
+ <span id="zahlSpan"></span>
  <vaadin-upload id="upload"></vaadin-upload>
 </vaadin-vertical-layout>
 </template>
